[lotus] Ensure undo data has correct block header hash
During reorgs, Fulcrum is failing consistency checks. It seems to
be because the header hash check is failing when loading the undo
data. Sure enough, this hash is stored as a standard BTC header
hash, so this would make sense. This commit fixes that issue --
although more issues may be hiding.